Revamp user experience with Add to Cart UI
By introducing the "Add to Cart" button with advanced functionality, we enhance the user experience and streamline the purchasing process. The unique cart item IDs enable precise tracking and management of individual products, even when multiple variations of the same product (such as color or size) are added to the cart.